Refactoring Convolution; factor out convolutions from layers.py
