+++ axlearn/common/conformer.py
@@ -30,9 +30,9 @@ from axlearn.common.attention import (
 )
 from axlearn.common.base_layer import BaseLayer
 from axlearn.common.config import REQUIRED, InstantiableConfig, Required, config_class
+from axlearn.common.convolution import Conv1D
 from axlearn.common.layers import (
     BatchNorm,
-    Conv1D,
     Dropout,
     GroupNorm,
     LayerNorm,
