+++ axlearn/vision/efficientdet.py
@@ -13,7 +13,8 @@ import numpy as np
 
 from axlearn.common.base_layer import BaseLayer
 from axlearn.common.config import REQUIRED, InstantiableConfig, Required, config_class
-from axlearn.common.layers import BatchNorm, Conv2D, get_activation_fn, set_norm_recursively
+from axlearn.common.convolution import Conv2D
+from axlearn.common.layers import BatchNorm, get_activation_fn, set_norm_recursively
 from axlearn.common.module import Module, Tensor, child_context
 from axlearn.common.param_init import (
     PARAM_REGEXP_BIAS,
