+++ axlearn/vision/mobilenets_blocks.py
@@ -28,7 +28,8 @@ from typing import Optional
 
 from axlearn.common.base_layer import BaseLayer
 from axlearn.common.config import REQUIRED, InstantiableConfig, Required, config_class
-from axlearn.common.layers import BatchNorm, Conv2D, get_activation_fn
+from axlearn.common.convolution import Conv2D
+from axlearn.common.layers import BatchNorm, get_activation_fn
 from axlearn.common.module import Module
 from axlearn.common.param_init import PerGroupInitializer
 from axlearn.common.utils import Tensor
