+++ axlearn/vision/resnet.py
@@ -33,9 +33,9 @@ from jax import numpy as jnp
 from axlearn.common import param_init
 from axlearn.common.base_layer import BaseLayer
 from axlearn.common.config import REQUIRED, InstantiableConfig, Required, config_class
+from axlearn.common.convolution import Conv2D
 from axlearn.common.layers import (
     BatchNorm,
-    Conv2D,
     SqueezeExcitation,
     StochasticDepth,
     get_activation_fn,
